Working unit tests and abs module paths
