+++ flytekit/core/tracker.py
@@ -1,8 +1,12 @@
+import importlib
 import importlib as _importlib
 import inspect
 import inspect as _inspect
-from typing import Callable
+import os
+import typing
+from typing import Callable, Tuple, Union
 
+from flytekit.configuration.feature_flags import FeatureFlags
 from flytekit.exceptions import system as _system_exceptions
 from flytekit.loggers import logger
 
@@ -55,7 +59,8 @@ class TrackedInstance(metaclass=InstanceTrackingMeta):
 
     @property
     def location(self) -> str:
-        return f"{self.instantiated_in}.{self.lhs}"
+        n, _, _ = extract_task_module(self)
+        return n
 
     @property
     def lhs(self):
@@ -163,3 +168,78 @@ def istestfunction(func) -> bool:
             mod_name = mod_name.split(".")[-1]
         return mod_name.startswith("test_")
     return False
+
+
+class _ModuleSanitizer(object):
+    """
+    Sanitizes and finds the absolute module path irrespective of the import location.
+    """
+
+    def __init__(self):
+        self._module_cache = {}
+
+    def _resolve_abs_module_name(self, basename: str, dirname: str, package_root: str) -> str:
+        """
+        Recursively finds the root python package under-which basename exists
+        """
+        # If we have already computed the module for this directory - return
+        if dirname in self._module_cache:
+            return self._module_cache[dirname]
+
+        # Let us remove any extensions like .py
+        basename = os.path.splitext(basename)[0]
+
+        if dirname == package_root:
+            return basename
+
+        # If we have reached a directory with no __init__, ignore
+        if "__init__.py" not in os.listdir(dirname):
+            return basename
+
+        # Now  recurse down such that we can extract the absolute module path
+        mod_name = self._resolve_abs_module_name(os.path.basename(dirname), os.path.dirname(dirname), package_root)
+        final_mod_name = f"{mod_name}.{basename}" if mod_name else basename
+        self._module_cache[dirname] = final_mod_name
+        return final_mod_name
+
+    def get_absolute_module_name(self, path: str, package_root: typing.Optional[str] = None) -> str:
+        """
+        Returns the absolute module path for a given python file path. This assumes that every module correctly contains
+        a __init__.py file. Absence of this file, indicates the root.
+        """
+        return self._resolve_abs_module_name(os.path.basename(path), os.path.dirname(path), package_root)
+
+
+_mod_sanitizer = _ModuleSanitizer()
+
+
+def extract_task_module(f: Union[Callable, TrackedInstance]) -> Tuple[str, str, str]:
+    """
+    Returns the task-name, absolute module and the string name of the callable.
+    :param f: A task or any other callable
+    :return: [name to use: str, module_name: str, function_name: str]
+    """
+
+    if isinstance(f, TrackedInstance):
+        mod = importlib.import_module(f.instantiated_in)
+        mod_name = mod.__name__
+        name = f.lhs
+    else:
+        mod = inspect.getmodule(f)
+        if mod is None:
+            raise AssertionError(f"Unable to determine module of {f}")
+        mod_name = mod.__name__
+        name = f.__name__.split(".")[-1]
+
+    if mod_name == "__main__":
+        return name, "", name
+
+    if FeatureFlags.FLYTE_PYTHON_PACKAGE_ROOT != ".":
+        package_root = (
+            FeatureFlags.FLYTE_PYTHON_PACKAGE_ROOT if FeatureFlags.FLYTE_PYTHON_PACKAGE_ROOT != "auto" else None
+        )
+        new_mod_name = _mod_sanitizer.get_absolute_module_name(inspect.getabsfile(f), package_root)
+        # We only replace the mod_name if it is more specific, else we already have a fully resolved path
+        if len(new_mod_name) > len(mod_name):
+            mod_name = new_mod_name
+    return f"{mod_name}.{name}", mod_name, name
