+++ tests/flytekit/unit/core/test_resolver.py
@@ -22,32 +22,29 @@ serialization_settings = flytekit.configuration.SerializationSettings(
 )
 
 
+@workflow
+def my_wf(a: int, b: str) -> typing.Tuple[int, str]:
+    @task
+    def t1(a: int) -> typing.Tuple[int, str]:
+        return a + 2, "world"
+
+    @task
+    def t2(a: str, b: str) -> str:
+        return b + a
+
+    x, y = t1(a=a)
+    d = t2(a=y, b=b)
+    return x, d
+
+
 def test_wf_resolving():
-    @workflow
-    def my_wf(a: int, b: str) -> typing.Tuple[int, str]:
-        @task
-        def t1(a: int) -> typing.Tuple[int, str]:
-            return a + 2, "world"
-
-        @task
-        def t2(a: str, b: str) -> str:
-            return b + a
-
-        x, y = t1(a=a)
-        d = t2(a=y, b=b)
-        return x, d
 
     x = my_wf(a=3, b="hello")
     assert x == (5, "helloworld")
 
     # Because the workflow is nested inside a test, calling location will fail as it tries to find the LHS that the
     # workflow was assigned to
-    with pytest.raises(Exception):
-        _ = my_wf.location
-
-    # Pretend my_wf was not actually nested, but somehow assigned to example_var_name at the module layer
-    my_wf._instantiated_in = "example.module"
-    my_wf._lhs = "example_var_name"
+    assert my_wf.location == "test_resolver.my_wf"
 
     workflows_tasks = my_wf.get_all_tasks()
     assert len(workflows_tasks) == 2  # Two tasks were declared inside
@@ -57,7 +54,7 @@ def test_wf_resolving():
     srz_t0_spec = get_serializable(OrderedDict(), serialization_settings, workflows_tasks[0])
     assert srz_t0_spec.template.container.args[-4:] == [
         "--resolver",
-        "example.module.example_var_name",
+        "test_resolver.my_wf",
         "--",
         "0",
     ]
@@ -65,7 +62,7 @@ def test_wf_resolving():
     srz_t1_spec = get_serializable(OrderedDict(), serialization_settings, workflows_tasks[1])
     assert srz_t1_spec.template.container.args[-4:] == [
         "--resolver",
-        "example.module.example_var_name",
+        "test_resolver.my_wf",
         "--",
         "1",
     ]
